Break out All/Fewest-line maps out of axialmap.h
