+++ salalib/axialmap.cpp
@@ -18,6 +18,9 @@
 
 // This is my code to make a set of axial lines from a set of boundary lines
 
+#include "salalib/alllinemap.h"
+#include "salalib/tidylines.h"
+#include "salalib/tolerances.h"
 #include <math.h>
 #include <float.h>
 #include <time.h>
@@ -38,20 +41,8 @@
 #define _finite finite
 #endif
 
-static const double TOLERANCE_A = 1e-9;
-static const double TOLERANCE_B = 1e-12;
-static const double TOLERANCE_C = 1e-6;
-
 ////////////////////////////////////////////////////////////////////////////////////////////
 
-static int compareValueTriplet(const void *p1, const void *p2)
-{
-   ValueTriplet *vp1 = (ValueTriplet *) p1;
-   ValueTriplet *vp2 = (ValueTriplet *) p2;
-   int v = vp1->value1 - vp2->value1;
-   return (vp1->value1 > vp2->value1 ? 1 : vp1->value1 < vp2->value1 ? -1 :
-          (vp1->value2 > vp2->value2 ? 1 : vp1->value2 < vp2->value2 ? -1 : 0));
-}
 
 static std::string makeFloatRadiusText(double radius)
 {
@@ -69,1152 +60,6 @@ static std::string makeFloatRadiusText(double radius)
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////
-
-AxialPolygons::AxialPolygons()
-{
-   m_pixel_polys = NULL;
-}
-
-AxialPolygons::~AxialPolygons()
-{
-   if (m_pixel_polys) {
-      for (int i = 0; i < m_cols; i++) {
-         delete [] m_pixel_polys[i];
-      }
-      delete [] m_pixel_polys;
-      m_pixel_polys = NULL;
-   }
-}
-
-AxialVertex AxialPolygons::makeVertex(const AxialVertexKey& vertexkey, const Point2f& openspace)
-{
-   auto vertPossIter = depthmapX::getMapAtIndex(m_vertex_possibles, vertexkey.m_ref_key);
-   AxialVertex av(vertexkey, vertPossIter->first, openspace);
-
-   // n.b., at this point, vertex key m_a and m_b are unfixed
-   pqvector<Point2f>& pointlist = vertPossIter->second;
-   if (pointlist.size() < 2) {
-      return av;
-   }
-
-   Point2f o = av.m_point - av.m_openspace;
-
-   // using an anglemap means that there are now no anti-clockwise vertices...
-   // TODO: (CS) Double as key is problematic - books have been written about double equality...
-   std::map<double,int> anglemap;
-   for (size_t i = 0; i < pointlist.size(); i++) {
-      anglemap.insert(std::make_pair( angle(openspace,av.m_point,pointlist[i]), i ));
-   }
-
-   av.m_ref_a = anglemap.begin()->second;
-   // TODO: is this supposed to be av.m_ref_b?
-   av.m_ref_a = anglemap.rbegin()->second;
-   Point2f a = av.m_point - pointlist.at( anglemap.begin()->second );
-   Point2f b = pointlist.at( anglemap.rbegin()->second ) - av.m_point;
-   av.m_a = a;
-   av.m_b = b;
-   a.normalise();
-   b.normalise();
-
-   double oa = det(o,a);
-   double ob = det(o,b);
-   double ab = det(a,b);
-
-   // can't handle these cases
-   if (fabs(oa) < TOLERANCE_A || fabs(ob) < TOLERANCE_A || fabs(ab) < TOLERANCE_A) {
-      // although note that if ab == 0 and you've already checked intersection, it can't be convex
-      return av;
-   }
-
-   // ADDED 4-Nov-04 -- In order to stop too many lines being generated, don't include
-   // points that do not change surface direction:  -- notice: will create problems with circles
-   if (fabs(dot(a,b)) > 0.999) {
-      return av;
-   }
-
-
-   if (sgn(oa) == sgn(ob)) {
-      // headon collision
-      if (sgn(oa) == 1) {
-         if (sgn(ab) == 1) {
-            // convex clockwise
-            av.m_convex = true;
-            av.m_clockwise = true;
-            av.m_axial = true;
-         }
-         else {
-            // n.b., these are turned away for axial formation
-            // concave clockwise
-            av.m_convex = false;
-            av.m_clockwise = true;
-            av.m_axial = false;
-         }
-      }
-   }
-   else {
-      // glancing blow
-      // concave clockwise
-      av.m_convex = false;
-      av.m_clockwise = true;
-      av.m_axial = true;
-   }
-
-   av.m_initialised = true;
-
-   return av;
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////////
-
-void AxialPolygons::clear()
-{
-   // clear any existing data
-   if (m_pixel_polys) {
-      for (int i = 0; i < m_cols; i++) {
-         delete [] m_pixel_polys[i];
-      }
-      delete [] m_pixel_polys;
-      m_pixel_polys = NULL;
-   }
-
-   m_vertex_possibles.clear();
-   m_vertex_polys.clear();
-   m_handled_list.clear();
-}
-
-void AxialPolygons::init(std::vector<Line>& lines, const QtRegion& region)
-{
-   // init pixelbase members
-   m_region = region;
-
-   // now tidy
-   TidyLines tidier;
-   tidier.tidy(lines, m_region);
-
-   // for easier debugging, the axial code is reused to make segments
-   ShapeGraph firstpass;
-   firstpass.init(lines.size(),m_region); // used to be double density
-   firstpass.initialiseAttributesAxial();
-   size_t i;
-   for (i = 0; i < lines.size(); i++) {
-      firstpass.makeLineShape(lines[i]);
-   }
-   firstpass.makeConnections();
-
-   lines.clear();
-   prefvec<Connector> connectionset;
-
-   // interesting... 1.0 may or may not work as intended
-   firstpass.makeSegmentMap(lines, connectionset, 1.0);
-
-   // now we have a set of lines and a set of connections...
-   // ...for the second pass, a bit of retro fitting to my original code is
-   // required
-   makeVertexPossibles(lines, connectionset);
-
-   initLines(lines.size(), m_region.bottom_left, m_region.top_right, 2);
-   // need to init before making pixel polys...
-   makePixelPolys();
-   // now also add lines
-   for (auto vertexPoss: m_vertex_possibles) {
-      for (size_t j = 0; j < vertexPoss.second.size(); j++) {
-         addLine(Line(vertexPoss.first,vertexPoss.second.at(j)));
-      }
-   }
-   sortPixelLines();
-}
-
-void AxialPolygons::makeVertexPossibles(const std::vector<Line>& lines, const prefvec<Connector>& connectionset)
-{
-   m_vertex_possibles.clear();
-   m_vertex_polys.clear();
-   int currpoly = -1;
-
-   size_t i = 0;
-
-   // TODO: (CS) these should be vectors, not raw pointers.
-   int *found[2];
-   found[0] = new int [lines.size()];
-   found[1] = new int [lines.size()];
-   for (i = 0; i < lines.size(); i++) {
-      found[0][i] = -1;
-      found[1][i] = -1;
-   }
-   pqvector<Point2f> pointlookup;
-   // three pass operation: (1) stack the lines
-   for (i = 0; i < lines.size(); i++) {
-      if (found[0][i] == -1) {
-         pointlookup.push_back(lines[i].start());
-         m_vertex_possibles.insert(std::make_pair(pointlookup.tail(),pqvector<Point2f>()));
-         m_vertex_polys.push_back(-1); // <- n.b., dummy entry for now, maintain with vertex possibles
-         found[0][i] = pointlookup.size() - 1;
-         for (size_t j = 0; j < connectionset[i].m_back_segconns.size(); j++) {
-            const SegmentRef& segref = connectionset[i].m_back_segconns.key(j);
-            int forwback = (segref.dir == 1) ? 0 : 1;
-            found[forwback][segref.ref] = found[0][i];
-         }
-      }
-      if (found[1][i] == -1) {
-         pointlookup.push_back(lines[i].end());
-         m_vertex_possibles.insert(std::make_pair(pointlookup.tail(),pqvector<Point2f>()));
-         m_vertex_polys.push_back(-1); // <- n.b., dummy entry for now, maintain with vertex possibles
-         found[1][i] = pointlookup.size() - 1;
-         for (size_t j = 0; j < connectionset[i].m_forward_segconns.size(); j++) {
-            const SegmentRef& segref = connectionset[i].m_forward_segconns.key(j);
-            int forwback = (segref.dir == 1) ? 0 : 1;
-            found[forwback][segref.ref] = found[1][i];
-         }
-      }
-   }
-   // three pass operation: (2) connect up vertex possibles
-   for (i = 0; i < lines.size(); i++) {
-      if (found[0][i] == -1 || found[1][i] == -1) {
-         // TODO: (CS) What are these integers being thrown?!
-         throw 1;
-      }
-      auto index0 = m_vertex_possibles.find(pointlookup.at(found[0][i]));
-      auto index1 = m_vertex_possibles.find(pointlookup.at(found[1][i]));
-      if (index0 == m_vertex_possibles.end() || index1 == m_vertex_possibles.end()) {
-         // TODO: (CS) What are these integers being thrown?!
-         throw 2;
-      }
-      index0->second.add(pointlookup.at(found[1][i]));
-      index1->second.add(pointlookup.at(found[0][i]));
-   }
-   delete [] found[0];
-   delete [] found[1];
-   // three pass operation: (3) create vertex poly entries
-   int current_poly = -1;
-   for (i = 0; i < m_vertex_possibles.size(); i++) {
-      if (m_vertex_polys[i] == -1) {
-         current_poly++;
-         pvecint addlist;
-         addlist.push_back(i);
-         while (addlist.size()) {
-            m_vertex_polys[addlist.tail()] = current_poly;
-            pqvector<Point2f>& connections = depthmapX::getMapAtIndex(m_vertex_possibles, addlist.tail())->second;
-            addlist.pop_back();
-            for (size_t j = 0; j < connections.size(); j++) {
-               int index = depthmapX::findIndexFromKey(m_vertex_possibles, connections[j]);
-               if (index == -1) {
-                  throw 3;
-               }
-               if (m_vertex_polys[index] == -1) {
-                  addlist.push_back(index);
-               }
-            }
-         }
-      }
-   }
-}
-
-void AxialPolygons::makePixelPolys()
-{
-   int i = 0;
-
-   // record all of this onto the pixel polygons
-   if (m_pixel_polys)
-   {
-      for (i = 0; i < m_cols; i++) {
-         delete [] m_pixel_polys[i];
-      }
-      delete [] m_pixel_polys;
-      m_pixel_polys = NULL;
-   }
-   m_pixel_polys = new pvecint *[m_cols];
-   for (i = 0; i < m_cols; i++) {
-      m_pixel_polys[i] = new pvecint[m_rows];
-   }
-   // now register the vertices in each pixel...
-   int j = -1;
-   for (auto vertPoss: m_vertex_possibles) {
-      j++;
-      PixelRef pix = pixelate(vertPoss.first);
-      m_pixel_polys[pix.x][pix.y].push_back(j);
-   }
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////////
-
-// almost identical to original!
-
-AxialVertexKey AxialPolygons::seedVertex(const Point2f& seed)
-{
-   AxialVertexKey seedvertex = NoVertex;
-   PixelRef seedref = pixelate(seed);
-   bool foundvertex = false;
-   // for spiralling outwards to find a vertex:
-   int dir = PixelRef::HORIZONTAL;
-   int sidelength = 1;
-   int runlength = 0;
-   int allboundaries = 0;
-
-   while (!foundvertex) {
-      for (size_t i = 0; i < m_pixel_polys[seedref.x][seedref.y].size(); i++) {
-         int vertexref = m_pixel_polys[seedref.x][seedref.y][i];
-         const Point2f& trialpoint = depthmapX::getMapAtIndex(m_vertex_possibles, vertexref)->first;
-         if (!intersect_exclude(Line(seed,trialpoint))) {
-            // yay... ...but wait... we need to see if it's a proper polygon vertex first...
-            seedvertex = vertexref;
-            foundvertex = true;
-         }
-      }
-      if (!foundvertex) {
-         seedref = seedref.move(dir);
-         // spiral outwards:
-         if (++runlength == sidelength) {
-            switch (dir) {
-            case PixelRef::HORIZONTAL:
-               dir = PixelRef::VERTICAL; runlength = 0;
-               break;
-            case PixelRef::VERTICAL:
-               dir = PixelRef::NEGHORIZONTAL; runlength = 0; sidelength++;
-               break;
-            case PixelRef::NEGHORIZONTAL:
-               dir = PixelRef::NEGVERTICAL; runlength = 0;
-               break;
-            case PixelRef::NEGVERTICAL:
-               dir = PixelRef::HORIZONTAL; runlength = 0; sidelength++;
-               break;
-            }
-         }
-         // check to make sure not off edge of system:
-         if (seedref.x < 0) {
-            allboundaries |= 0x01; seedref.x = 0;
-         }
-         if (seedref.y < 0) {
-            allboundaries |= 0x02; seedref.y = 0;
-         }
-         if (seedref.x >= m_cols) {
-            allboundaries |= 0x04; seedref.x = m_cols - 1;
-         }
-         if (seedref.y >= m_rows) {
-            allboundaries |= 0x08; seedref.y = m_rows - 1;
-         }
-         if (allboundaries == 0x0f) {
-            return NoVertex;
-         }
-      }
-   }
-   return seedvertex;
-}
-
-// adds any axial lines from this point to the list of lines, adds any unhandled visible vertices it finds to the openvertices list
-// axial lines themselves are added to the lines list - the axial line is only there to record the key vertices that comprise the line
-void AxialPolygons::makeAxialLines(pqvector<AxialVertex>& openvertices, prefvec<Line>& lines, prefvec<pvecint>& keyvertices, prefvec<PolyConnector>& poly_connections, pqvector<RadialLine>& radial_lines)
-{
-   AxialVertex vertex = openvertices.tail();
-   openvertices.pop_back();
-
-   m_handled_list.add(vertex);
-
-   int i = -1;
-   for (auto vertPoss: m_vertex_possibles) {
-      i++;
-      if (i == vertex.m_ref_key) {
-         continue;
-      }
-      bool possible = false, stubpossible = false;
-      Point2f p = vertPoss.first - vertex.m_point;
-      if (vertex.m_convex) {
-         if (det(vertex.m_a,p) > 0 && det(vertex.m_b,p) > 0) {
-            possible = true;
-         }
-      }
-      else {
-         // left of b and right of a or left of a and right of b
-         if (det(p,vertex.m_a) * det(p,vertex.m_b) < 0) {
-            possible = true;
-         }
-         else if (det(p,vertex.m_a) < TOLERANCE_A && det(p,vertex.m_b) < TOLERANCE_A) {
-            stubpossible = true;
-         }
-      }
-      if (possible || stubpossible) {
-         Line line(vertPoss.first,vertex.m_point);
-         if (!intersect_exclude(line)) {
-            AxialVertex next_vertex = makeVertex(AxialVertexKey(i),vertex.m_point);
-            if (next_vertex.m_initialised && m_handled_list.searchindex(next_vertex) == paftl::npos) {
-               openvertices.add(next_vertex); // <- note, add ignores duplicate adds (each vertex tends to be added multiple times before this vertex is handled itself)
-               bool shortline_segend = false;
-               Line shortline = line;
-               if (!vertex.m_convex && possible) {
-                  Line ext(line.t_end(), line.t_end() + (line.t_end() - line.t_start()));
-                  ext.ray(1, m_region);
-                  cutLine(ext, 1);
-                  line = Line(line.t_start(), ext.t_end());
-                  // for radial line segend calc:
-                  if (det(-p,vertex.m_b) < 0) {
-                     shortline_segend = true;
-                  }
-               }
-               if (m_vertex_polys[vertex.m_ref_key] != m_vertex_polys[next_vertex.m_ref_key]) { // must be on separate polygons
-                  // radial line(s) (for new point)
-                  RadialLine radialshort(next_vertex, shortline_segend, vertex.m_point,next_vertex.m_point,next_vertex.m_point+next_vertex.m_b);
-                  poly_connections.push_back( PolyConnector(shortline, (RadialKey)radialshort) );
-                  radial_lines.add(radialshort);
-                  if (!vertex.m_convex && possible) {
-                     Line longline = Line(vertPoss.first,line.t_end());
-                     RadialLine radiallong(radialshort);
-                     radiallong.segend = shortline_segend ? 0 : 1;
-                     poly_connections.push_back( PolyConnector(longline, (RadialKey)radiallong) );
-                     radial_lines.add(radiallong);
-                  }
-               }
-               shortline_segend = false;
-               if (!next_vertex.m_convex && next_vertex.m_axial) {
-                  Line ext(line.t_start() - (line.t_end() - line.t_start()), line.t_start());
-                  ext.ray(0, m_region);
-                  cutLine(ext, 0);
-                  line = Line(ext.t_start(), line.t_end());
-                  // for radial line segend calc:
-                  if (det(p,next_vertex.m_b) < 0) {
-                     shortline_segend = true;
-                  }
-               }
-               if (m_vertex_polys[vertex.m_ref_key] != m_vertex_polys[next_vertex.m_ref_key]) { // must be on separate polygons
-                  // radial line(s) (for original point)
-                  RadialLine radialshort(vertex, shortline_segend, next_vertex.m_point,vertex.m_point,vertex.m_point+vertex.m_b);
-                  poly_connections.push_back( PolyConnector(shortline, (RadialKey)radialshort) );
-                  radial_lines.add(radialshort);
-                  if (!next_vertex.m_convex && next_vertex.m_axial) {
-                     Line longline = Line(line.t_start(),vertex.m_point);
-                     RadialLine radiallong(radialshort);
-                     radiallong.segend = shortline_segend ? 0 : 1;
-                     poly_connections.push_back( PolyConnector(longline, (RadialKey)radiallong) );
-                     radial_lines.add(radiallong);
-                  }
-               }
-               if (possible && next_vertex.m_axial) {
-                  // axial line
-                  lines.push_back(line);
-                  keyvertices.push_back(pvecint());
-                  if (vertex.m_convex) {
-                     keyvertices.tail().add(vertex.m_ref_key);
-                  }
-                  if (next_vertex.m_convex) {
-                     keyvertices.tail().add(next_vertex.m_ref_key);
-                  }
-               }
-            }
-         }
-      }
-   }
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-
-// not really used as yet, a feature to make all the polygons from the vertex
-// possibles list
-
-void AxialPolygons::makePolygons(prefvec<pqvector<Point2f>>& polygons)
-{
-   prefvec<pvecint> handled_list;
-   for (size_t j = 0; j < m_vertex_possibles.size(); j++) {
-      handled_list.push_back(pvecint());
-   }
-
-   int i = -1;
-   for (auto vertPoss: m_vertex_possibles) {
-      i++;
-      if (vertPoss.second.size() == 1) {
-         continue;
-      }
-      for (size_t j = 0; j < vertPoss.second.size(); j++) {
-         if (handled_list[i].findindex(j) != paftl::npos) {
-            continue;
-         }
-         handled_list[i].push_back(j);
-         const Point2f& key = vertPoss.first;
-         pqvector<Point2f> polygon;
-         polygon.push_back(key);
-         Point2f curr = vertPoss.second.at(j);
-         Point2f last = key;
-         bool good = true;
-         while (curr != key) {
-            auto vertPossIter = m_vertex_possibles.find(curr);
-            polygon.push_back(curr);
-            // hunt down left most
-            int winner = -1, wayback = -1;
-            double minangle = 2 * M_PI;
-            for (size_t k = 0; k < vertPossIter->second.size(); k++) {
-               Point2f next = vertPossIter->second.at(k);
-               if (last != next) {
-                  double thisangle = angle(last,curr,next);
-                  if (thisangle < minangle) {
-                     // check not going to a dead end:
-                     if (m_vertex_possibles.find(vertPossIter->second.at(k))->second.size() > 1) {
-                        minangle = thisangle;
-                        winner = k;
-                     }
-                  }
-               }
-               else {
-                  wayback = k;
-               }
-            }
-            if (winner == -1) {
-               // this happens when you follow a false trail -- go back the way you came!
-               winner = wayback;
-            }
-            handled_list[std::distance(m_vertex_possibles.begin(), vertPossIter)].push_back(winner);
-            last = curr;
-            curr = vertPossIter->second.at(winner);
-         }
-         if (good) {
-            polygons.push_back(polygon);
-         }
-         good = true;
-      }
-   }
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-
-bool RadialLine::cuts(const Line& l) const
-{
-   if (fabs(det(l.end() - keyvertex,l.end() - l.start())) < TOLERANCE_A) {
-      // point on line, check that openspace and next vertex are on opposite sides of the line
-      Point2f x = l.end() - keyvertex;
-      Point2f y = nextvertex - keyvertex;
-      Point2f z = openspace - keyvertex;
-      x.normalise();
-      y.normalise();
-      z.normalise();
-      if (sgn(det(x, y)) == sgn(det(x, z)) && fabs(det(x, z)) > TOLERANCE_A) {
-         return false;
-      }
-   }
-   // keyvertex not on line... the line's been cut:
-   return true;
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-
-// for this to work, your polygons should not intersect each other!
-
-bool ShapeGraphs::makeAllLineMap(Communicator *comm, std::vector<SpacePixelFile> &drawingLayers, const Point2f& seed)
-{
-   if (comm) {
-      comm->CommPostMessage( Communicator::NUM_STEPS, 3 );
-      comm->CommPostMessage( Communicator::CURRENT_STEP, 1 );
-   }
-
-   // this has a nasty habit of crashing if reused...
-   // reset everything at the top level, including any existing all-line map:
-   m_polygons.clear();
-   m_poly_connections.clear();
-   m_radial_lines.clear();
-
-   // this is an index to look up the all line map, used by UI to determine if can make fewest line map
-   // note: it is not saved for historical reasons
-   if (m_all_line_map != -1) {
-      removeMap(m_all_line_map);
-      m_all_line_map = -1;
-   }
-
-   // starting off... finding a polygon...
-   // for ease, I'm just going to make a construction line set from all the visible lines...
-
-   QtRegion region;
-   int size = 0;
-
-   std::vector<Line> lines;
-
-   // add all visible layers to the set of polygon lines...
-   for (const auto& pixelGroup: drawingLayers) {
-      for (const auto& pixel: pixelGroup.m_spacePixels) {
-         if (pixel.isShown()) {
-            if (region.atZero()) {
-               region = pixel.getRegion();
-            }
-            else {
-               region = runion(region, pixel.getRegion());
-            }
-            std::vector<SimpleLine> newLines = pixel.getAllShapesAsLines();
-            for (const auto& line: newLines) {
-               lines.push_back(Line(line.start(), line.end()));
-            }
-         }
-      }
-   }
-
-   region.grow(1.30);
-   m_polygons.init(lines, region);
-   m_polygons.m_handled_list.clear();
-
-   // find a corner visible from the seed:
-   AxialVertexKey seedvertex = m_polygons.seedVertex( seed );
-
-   if (seedvertex == NoVertex) {
-      // oops... can't find a visible vertex
-      return false;
-   }
-
-   // okay, we've got as far as finding a seed corner, now the real fun begins...
-   // test outwards from corner, add other corners to
-   // test set...
-   prefvec<Line> axiallines;
-   prefvec<pvecint> preaxialdata;
-   // also poly_connections used in fewest line axial map construction:
-   m_poly_connections.clear();
-   m_radial_lines.clear();
-
-   AxialVertex vertex = m_polygons.makeVertex(seedvertex, seed);
-   if (!vertex.m_initialised) {
-      // oops... can't init for some reason
-      return false;
-   }
-
-
-   time_t atime = 0;
-   int count = 0;
-   if (comm) {
-      qtimer( atime, 0 );
-      comm->CommPostMessage( Communicator::CURRENT_STEP, 2 );
-      comm->CommPostMessage( Communicator::NUM_RECORDS, m_polygons.m_vertex_possibles.size() );
-   }
-
-   pqvector<AxialVertex> openvertices;
-   openvertices.add(vertex);
-   while (openvertices.size()) {
-      m_polygons.makeAxialLines(openvertices, axiallines, preaxialdata, m_poly_connections, m_radial_lines);
-      count++;
-      //
-      if (comm) {
-         if (qtimer( atime, 500 )) {
-            if (comm->IsCancelled()) {
-               throw Communicator::CancelledException();
-            }
-            comm->CommPostMessage( Communicator::CURRENT_RECORD, count );
-         }
-      }
-
-   }
-
-   if (comm) {
-      comm->CommPostMessage( Communicator::CURRENT_STEP, 3 );
-      comm->CommPostMessage( Communicator::CURRENT_RECORD, 0 );
-   }
-
-   // cut out duplicates:
-   int removed = 0;  // for testing purposes
-   for (size_t j = 0; j < axiallines.size(); j++) {
-      for (size_t k = axiallines.size() - 1; k > j; k--) {
-         double maxdim = __max(region.width(),region.height());
-         if (approxeq(axiallines[j].start(), axiallines[k].start(), maxdim * TOLERANCE_B) && approxeq(axiallines[j].end(), axiallines[k].end(), maxdim * TOLERANCE_B)) {
-            for (size_t m = 0; m < preaxialdata[k].size(); m++) {
-               preaxialdata[j].add(preaxialdata[k][m]);
-            }
-            preaxialdata.remove_at(k);
-            axiallines.remove_at(k);
-            removed++;
-         }
-      }
-   }
-
-   // create the all line map layer...
-   m_all_line_map = addMap("All-Line Map", ShapeMap::ALLLINEMAP);
-
-   ShapeGraph& alllinemap = at(m_all_line_map);
-   // make sure it's cleared fully
-   alllinemap.clearAll();
-
-   region.grow(0.99); // <- this paired with crop code below to prevent error
-   alllinemap.init(axiallines.size(),m_polygons.m_region);  // used to be double density here
-   alllinemap.initialiseAttributesAxial();
-   for (size_t k = 0; k < axiallines.size(); k++) {
-      axiallines[k].crop(region); // <- should be cropped anyway, but causing an error
-      alllinemap.makeLineShape(axiallines[k]);
-   }
-
-   // n.b. make connections also initialises attributes
-   // -> don't know what this was for: alllinemap.sortBins(m_poly_connections);
-   alllinemap.makeConnections(preaxialdata);
-
-   alllinemap.m_keyvertexcount = m_polygons.m_vertex_possibles.size();
-
-   // we can stop here for all line axial map!
-   setDisplayedMapRef(m_all_line_map);
-
-   return true;
-}
-
-bool ShapeGraphs::makeFewestLineMap(Communicator *comm, bool replace_existing)
-{
-   // no all line map
-   if (m_all_line_map == -1) {
-      return false;
-   }
-
-   if (comm) {
-      comm->CommPostMessage( Communicator::NUM_STEPS, 2 );
-      comm->CommPostMessage( Communicator::CURRENT_STEP, 1 );
-   }
-
-   pafsrand((unsigned int)time(NULL));
-
-   // make one rld for each radial line...
-   std::map<RadialKey,pvecint> radialdivisions;
-   size_t i;
-   for (i = 0; i < m_radial_lines.size(); i++) {
-      radialdivisions.insert(std::make_pair( (RadialKey) m_radial_lines[i], pvecint() ));
-   }
-
-   // also, a list of radial lines cut by each axial line
-   std::map<int,pvecint> ax_radial_cuts;
-   std::map<int,pvecint> ax_seg_cuts;
-   for (auto shape: at(m_all_line_map).m_shapes) {
-      ax_radial_cuts.insert(std::make_pair(shape.first, pvecint()));
-      ax_seg_cuts.insert(std::make_pair(shape.first, pvecint()));
-   }
-
-   // make divisions -- this is the slow part and the comm updates
-   at(m_all_line_map).makeDivisions(m_poly_connections, m_radial_lines, radialdivisions, ax_radial_cuts, comm);
-
-   // the slow part is over, we're into the final straight... reset the current record flag:
-   if (comm) {
-      comm->CommPostMessage( Communicator::CURRENT_STEP, 2 );
-      comm->CommPostMessage( Communicator::CURRENT_RECORD, 0 );
-   }
-
-   // a little further setting up is still required...
-   std::map<RadialKey,RadialSegment> radialsegs;
-
-   // now make radial segments from the radial lines... (note, start at 1)
-   for (i = 1; i < m_radial_lines.size(); i++) {
-      if (m_radial_lines[i].vertex == m_radial_lines[i-1].vertex) {
-         if (m_radial_lines[i].ang == m_radial_lines[i-1].ang) {
-            continue;
-         }
-         else {
-            radialsegs.insert(std::make_pair( (RadialKey)m_radial_lines[i], (RadialSegment)m_radial_lines[i-1]));
-         }
-      }
-   }
-
-   // and segment divisors from the axial lines...
-   // TODO: (CS) Restructure this to get rid of all those brittle parallel data structure
-   auto axIter = ax_radial_cuts.begin();
-   auto axSeg = ax_seg_cuts.begin();
-   for (i = 0; i < at(m_all_line_map).m_shapes.size(); i++) {
-      for (size_t j = 1; j < axIter->second.size(); j++) {
-         // note similarity to loop above
-         RadialKey rk_end = m_radial_lines[axIter->second[j]];
-         RadialKey rk_start = m_radial_lines[axIter->second[j-1]];
-         if (rk_start.vertex == rk_end.vertex) {
-            auto radialSegIter = radialsegs.find(rk_end);
-            if (radialSegIter != radialsegs.end() && rk_start == radialSegIter->second.radial_b) {
-               radialSegIter->second.add(axIter->first);
-               axSeg->second.add(std::distance(radialsegs.begin(), radialSegIter));
-            }
-         }
-      }
-      axIter++;
-      axSeg++;
-   }
-
-   // and a little more setting up: key vertex relationships
-   prefvec<pvecint> keyvertexconns;
-   int *keyvertexcounts = new int [at(m_all_line_map).m_keyvertexcount];
-   for (int x = 0; x < at(m_all_line_map).m_keyvertexcount; x++) {
-      keyvertexcounts[x] = 0;
-   }
-   // this sets up a two step relationship: looks for the key vertices for all lines connected to you
-   for (size_t y = 0; y < at(m_all_line_map).m_connectors.size(); y++) {
-      keyvertexconns.push_back(pvecint());
-      Connector& axa = at(m_all_line_map).m_connectors[y];
-      for (size_t z = 0; z < axa.m_connections.size(); z++) {
-         pvecint& axb = at(m_all_line_map).m_keyvertices[axa.m_connections[z]];
-         for (size_t zz = 0; zz < axb.size(); zz++) {
-            if (keyvertexconns[y].searchindex(axb[zz]) == paftl::npos) {
-               keyvertexconns[y].add(axb[zz],paftl::ADD_HERE);
-               keyvertexcounts[axb[zz]] += 1;
-            }
-         }
-      }
-   }
-
-   // ok, after this fairly tedious set up, we are ready to go...
-   // note axradialcuts aren't required anymore...
-
-   AxialMinimiser minimiser(at(m_all_line_map), ax_seg_cuts.size(), radialsegs.size());
-
-   prefvec<Line> lines_s, lines_m;
-
-   minimiser.removeSubsets(ax_seg_cuts, radialsegs, radialdivisions, m_radial_lines, keyvertexconns, keyvertexcounts);
-
-   // make new lines here (assumes line map has only lines)
-   int k = -1;
-   for (auto& shape: at(m_all_line_map).m_shapes) {
-      k++;
-      if (!minimiser.removed(k)) {
-         lines_s.push_back( shape.second.getLine() );
-      }
-   }
-
-   minimiser.fewestLongest(ax_seg_cuts, radialsegs, radialdivisions, m_radial_lines, keyvertexconns, keyvertexcounts);
-
-   // make new lines here (assumes line map has only lines
-   for (k = 0; k < at(m_all_line_map).m_shapes.size(); k++) {
-      if (!minimiser.removed(k)) {
-         lines_m.push_back( depthmapX::getMapAtIndex(at(m_all_line_map).m_shapes, k)->second.getLine() );
-      }
-   }
-
-   delete [] keyvertexcounts;
-
-   int subsetmapindex = getMapRef("Fewest-Line Map (Subsets)");
-   if (subsetmapindex == -1) {
-      // didn't used to have hyphenation, try once more:
-      subsetmapindex = getMapRef("Fewest Line Map (Subsets)");
-   }
-   if (subsetmapindex == -1) {
-      // create the fewest line map layer...
-      subsetmapindex = addMap("Fewest-Line Map (Subsets)",ShapeMap::AXIALMAP);
-      // note: new map has replace_existing set to true to ensure "init"
-      replace_existing = true;
-   }
-   ShapeGraph& fewestlinemap_subsets = at(subsetmapindex);
-   // note: new map has replace_existing set to true to ensure "init"
-   if (replace_existing) {
-      fewestlinemap_subsets.clearAll();
-      fewestlinemap_subsets.init(lines_s.size(),m_polygons.m_region); // used to have a '2' for double pixel density
-   }
-   fewestlinemap_subsets.initialiseAttributesAxial();
-   for (k = 0; k < lines_s.size(); k++) {
-      fewestlinemap_subsets.makeLineShape(lines_s[k]);
-   }
-   fewestlinemap_subsets.makeConnections();
-
-   int minimalmapindex = getMapRef("Fewest-Line Map (Minimal)");
-   if (minimalmapindex == -1) {
-      // didn't used to have hyphenation, try once more:
-      minimalmapindex = getMapRef("Fewest Line Map (Minimal)");
-   }
-   if (minimalmapindex == -1) {
-      // create the fewest line map layer...
-      minimalmapindex = addMap("Fewest-Line Map (Minimal)",ShapeMap::AXIALMAP);
-      // note: new map has replace_existing set to true to ensure "init"
-      replace_existing = true;
-   }
-   ShapeGraph& fewestlinemap_minimal = at(minimalmapindex);
-   // note: new map has replace_existing set to true to ensure "init"
-   if (replace_existing) {
-      fewestlinemap_minimal.clearAll();
-      fewestlinemap_minimal.init(lines_m.size(),m_polygons.m_region); // used to have a '2' for double pixel density
-   }
-   fewestlinemap_minimal.initialiseAttributesAxial();
-   for (k = 0; k < lines_m.size(); k++) {
-      fewestlinemap_minimal.makeLineShape(lines_m[k]);
-   }
-   fewestlinemap_minimal.makeConnections();
-
-   setDisplayedMapRef(subsetmapindex);
-
-   return true;
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-
-AxialMinimiser::AxialMinimiser(const ShapeGraph& alllinemap, int no_of_axsegcuts, int no_of_radialsegs)
-{
-   m_alllinemap = (ShapeGraph *) &alllinemap;
-
-   m_vps = new ValueTriplet[no_of_axsegcuts];
-   m_removed = new bool [no_of_axsegcuts];
-   m_affected = new bool [no_of_axsegcuts];
-   m_vital = new bool [no_of_axsegcuts];
-   m_radialsegcounts = new int [no_of_radialsegs];
-}
-
-AxialMinimiser::~AxialMinimiser()
-{
-   delete [] m_vital;
-   delete [] m_affected;
-   delete [] m_radialsegcounts;
-   delete [] m_vps;
-   delete [] m_removed;
-}
-
-// Alan and Bill's algo...
-
-void AxialMinimiser::removeSubsets(std::map<int,pvecint>& axsegcuts, std::map<RadialKey,RadialSegment>& radialsegs, std::map<RadialKey,pvecint>& rlds,  pqvector<RadialLine>& radial_lines, prefvec<pvecint>& keyvertexconns, int *keyvertexcounts)
-{
-   bool removedflag = true;
-   int counterrors = 0;
-
-   m_axialconns = m_alllinemap->m_connectors;
-
-   for (size_t x = 0; x < radialsegs.size(); x++) {
-      m_radialsegcounts[x] = 0;
-   }
-   int y = -1;
-   for (auto axSegCut: axsegcuts) {
-      y++;
-      for (size_t z = 0; z < axSegCut.second.size(); z++) {
-         m_radialsegcounts[axSegCut.second[z]] += 1;
-      }
-      m_removed[y] = false;
-      m_vital[y] = false;
-      m_affected[y] = true;
-      m_vps[y].index = y;
-      double length = m_axialconns[y].m_connections.size();
-      m_vps[y].value1 = (int) length;
-      length = depthmapX::getMapAtIndex(m_alllinemap->m_shapes, y)->second.getLine().length();
-      m_vps[y].value2 = (float) length;
-   }
-
-   // sort according to number of connections then length
-   qsort(m_vps,m_axialconns.size(),sizeof(ValueTriplet),compareValueTriplet);
-
-   while (removedflag) {
-
-      removedflag = false;
-      for (size_t i = 0; i < m_axialconns.size(); i++) {
-         int ii = m_vps[i].index;
-         if (m_removed[ii] || !m_affected[ii] || m_vital[ii]) {
-            continue;
-         }
-         // vital connections code (uses original unaltered connections)
-         {
-            bool vitalconn = false;
-            for (size_t j = 0; j < keyvertexconns[ii].size(); j++) {
-               // first check to see if removing this line will cause elimination of a vital connection
-               if (keyvertexcounts[keyvertexconns[ii][j]] <= 1) {
-                  // connect vital... just go on to the next one:
-                  vitalconn = true;
-                  break;
-               }
-            }
-            if (vitalconn) {
-               m_vital[ii] = true;
-               continue;
-            }
-         }
-         //
-         Connector& axa = m_axialconns[ii];
-         m_affected[ii] = false;
-         bool subset = false;
-         for (size_t j = 0; j < axa.m_connections.size(); j++) {
-            int indextob = axa.m_connections[j];
-            if (indextob == ii || m_removed[indextob]) { // <- removed[indextob] should never happen as it should have been removed below
-               continue;
-            }
-            Connector& axb = m_axialconns[indextob];
-            if (axa.m_connections.size() <= axb.m_connections.size()) {
-               // change to 10.08, coconnecting is 1 -> connection to other line is implicitly handled
-               int coconnecting = 1;
-               // first check it's a connection subset
-               // note that changes in 10.08 mean that lines no longer connect to themselves
-               // this means that the subset 1 connects {2,3} and 2 connects {1,3} are equivalent
-               for (size_t axai = 0, axbi = 0; axai < axa.m_connections.size() && axbi < axb.m_connections.size(); axai++, axbi++) {
-                  // extra 10.08 -> step over connection to b
-                  if (axa.m_connections[axai] == indextob) {
-                     axai++;
-                  }
-                  // extra 10.08 add axb.m_connections[axbi] == ii -> step over connection to a
-                  while (axbi < axb.m_connections.size() && (axb.m_connections[axbi] == ii || axa.m_connections[axai] > axb.m_connections[axbi])) {
-                     axbi++;
-                  }
-                  if (axbi >= axb.m_connections.size()) {
-                     break;
-                  }
-                  else if (axa.m_connections[axai] == axb.m_connections[axbi]) {
-                     coconnecting++;
-                  }
-                  else if (axa.m_connections[axai] < axb.m_connections[axbi]) {
-                     break;
-                  }
-               }
-               if (coconnecting >= (int)axa.m_connections.size()) {
-                  subset = true;
-                  break;
-               }
-            }
-         }
-         if (subset) {
-            size_t removeindex = ii;
-            // now check removing it won't break any topological loops
-            bool presumedvital = false;
-            auto& axSegCut = depthmapX::getMapAtIndex(axsegcuts, removeindex)->second;
-            for (size_t k = 0; k < axSegCut.size(); k++) {
-               if (m_radialsegcounts[axSegCut[k]] <= 1) {
-                  presumedvital = true;
-                  break;
-               }
-            }
-            if (presumedvital) {
-               presumedvital = checkVital(removeindex,axSegCut,radialsegs,rlds,radial_lines);
-            }
-            if (presumedvital) {
-               m_vital[removeindex] = true;
-            }
-            // if not, remove it...
-            if (!m_vital[removeindex]) {
-               m_removed[removeindex] = true;
-               pvecint& affectedconnections = m_axialconns[removeindex].m_connections;
-               size_t k;
-               for (k = 0; k < affectedconnections.size(); k++) {
-                  if (!m_removed[affectedconnections[k]]) {
-                     pvecint& connections = m_axialconns[affectedconnections[k]].m_connections;
-                     size_t index = connections.searchindex(removeindex);
-                     if (index != paftl::npos) {
-                        connections.remove_at(index);
-                     }
-                     m_affected[affectedconnections[k]] = true;
-                  }
-               }
-               removedflag = true;
-               for (k = 0; k < axSegCut.size(); k++) {
-                  m_radialsegcounts[axSegCut[k]] -= 1;
-               }
-               // vital connections
-               for (k = 0; k < keyvertexconns[removeindex].size(); k++) {
-                  keyvertexcounts[keyvertexconns[removeindex][k]] -= 1;
-               }
-            }
-         }
-      }
-   }
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-
-// My algo... v. simple... fewest longest
-
-void AxialMinimiser::fewestLongest(std::map<int,pvecint>& axsegcuts, std::map<RadialKey,RadialSegment>& radialsegs, std::map<RadialKey, pvecint> &rlds, pqvector<RadialLine>& radial_lines, prefvec<pvecint>& keyvertexconns, int *keyvertexcounts)
-{
-   //m_axialconns = m_alllinemap->m_connectors;
-   int livecount = 0;
-
-   for (size_t y = 0; y < m_axialconns.size(); y++) {
-      if (!m_removed[y] && !m_vital[y]) {
-         m_vps[livecount].index = (int) y;
-         m_vps[livecount].value1 = (int) m_axialconns[y].m_connections.size();
-         m_vps[livecount].value2 = (float) depthmapX::getMapAtIndex(m_alllinemap->m_shapes, y)->second.getLine().length();
-         livecount++;
-      }
-   }
-
-   qsort(m_vps,livecount,sizeof(ValueTriplet),compareValueTriplet);
-
-   for (int i = 0; i < livecount; i++) {
-
-      int j = m_vps[i].index;
-      // vital connections code (uses original unaltered connections)
-      bool vitalconn = false;
-      size_t k;
-      for (k = 0; k < keyvertexconns[j].size(); k++) {
-         // first check to see if removing this line will cause elimination of a vital connection
-         if (keyvertexcounts[keyvertexconns[j][k]] <= 1) {
-            // connect vital... just go on to the next one:
-            vitalconn = true;
-            break;
-         }
-      }
-      if (vitalconn) {
-         continue;
-      }
-      //
-      bool presumedvital = false;
-      auto &axSegCut = depthmapX::getMapAtIndex(axsegcuts, j)->second;
-      for (k = 0; k < axSegCut.size(); k++) {
-         if (m_radialsegcounts[axSegCut[k]] <= 1) {
-            presumedvital = true;
-            break;
-         }
-      }
-      if (presumedvital) {
-         presumedvital = checkVital(j,axSegCut,radialsegs,rlds,radial_lines);
-      }
-      if (!presumedvital) {
-         // don't let anything this is connected to go down to zero connections
-         pvecint& affectedconnections = m_axialconns[j].m_connections;
-         for (size_t k = 0; k < affectedconnections.size(); k++) {
-            if (!m_removed[affectedconnections[k]]) {
-               pvecint& connections = m_axialconns[affectedconnections[k]].m_connections;
-               if (connections.size() <= 2) { // <- note number of connections includes itself... so you and one other
-                  presumedvital = true;
-                  break;
-               }
-            }
-         }
-      }
-      if (!presumedvital) {
-         m_removed[j] = true;
-         pvecint& affectedconnections = m_axialconns[j].m_connections;
-         size_t k;
-         for (k = 0; k < affectedconnections.size(); k++) {
-            if (!m_removed[affectedconnections[k]]) {
-               pvecint& connections = m_axialconns[affectedconnections[k]].m_connections;
-               size_t index = connections.searchindex(j);
-               if (index != paftl::npos) {
-                  connections.remove_at(index);
-               }
-               m_affected[affectedconnections[k]] = true;
-            }
-         }
-         for (k = 0; k < axSegCut.size(); k++) {
-            m_radialsegcounts[axSegCut[k]] -= 1;
-         }
-         // vital connections
-         for (k = 0; k < keyvertexconns[j].size(); k++) {
-            keyvertexcounts[keyvertexconns[j][k]] -= 1;
-         }
-      }
-   }
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
-
-bool AxialMinimiser::checkVital(int checkindex, pvecint& axsegcuts, std::map<RadialKey,RadialSegment>& radialsegs, std::map<RadialKey, pvecint> &rlds, pqvector<RadialLine>& radial_lines)
-{
-   std::map<int,SalaShape>& axiallines = m_alllinemap->m_shapes;
-
-   bool presumedvital = true;
-   int nonvitalcount = 0, vitalsegs = 0;
-   // again, this time more rigourously... check any connected pairs don't cover the link...
-   for (size_t k = 0; k < axsegcuts.size(); k++) {
-      if (m_radialsegcounts[axsegcuts[k]] <= 1) {
-         bool nonvitalseg = false;
-         vitalsegs++;
-         auto radialSegIter = depthmapX::getMapAtIndex(radialsegs, axsegcuts[k]);
-         const RadialKey& key = radialSegIter->first;
-         RadialSegment& seg = radialSegIter->second;
-         pvecint& divisorsa = rlds.find(key)->second;
-         pvecint& divisorsb = rlds.find(seg.radial_b)->second;
-         RadialLine& rlinea = radial_lines.search(key);
-         RadialLine& rlineb = radial_lines.search(seg.radial_b);
-         for (size_t divi = 0; divi < divisorsa.size(); divi++) {
-            if (divisorsa[divi] == checkindex || m_removed[divisorsa[divi]]) {
-               continue;
-            }
-            for (size_t divj = 0; divj < divisorsb.size(); divj++) {
-               if (divisorsb[divj] == checkindex || m_removed[divisorsb[divj]]) {
-                  continue;
-               }
-               if (m_axialconns[divisorsa[divi]].m_connections.searchindex(divisorsb[divj]) != paftl::npos) {
-                  // as a further challenge, they must link within in the zone of interest, not on the far side of it... arg!
-                  Point2f p = intersection_point(axiallines[divisorsa[divi]].getLine(),axiallines[divisorsb[divj]].getLine(),TOLERANCE_A);
-                  if (p.insegment(rlinea.keyvertex,rlinea.openspace,rlineb.openspace,TOLERANCE_A)) {
-                     nonvitalseg = true;
-                  }
-               }
-            }
-         }
-         if (nonvitalseg) {
-            nonvitalcount++;
-         }
-      }
-   }
-   if (nonvitalcount == vitalsegs) {
-      presumedvital = false;
-   }
-   return presumedvital;
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////
 
 // convert line layers to an axial map
 
@@ -1778,28 +623,6 @@ bool ShapeGraphs::read( std::istream& stream, int version )
 
    ShapeMaps<ShapeGraph>::read(stream,version);
 
-   // these are additional essentially for all line axial maps
-   // should probably be kept *with* the all line axial map...
-   m_poly_connections.clear();
-   m_poly_connections.read(stream);
-   m_radial_lines.clear();
-   m_radial_lines.read(stream);
-
-   // this is an index to look up the all line map, used by UI to determine if can make fewest line map
-   // note: it is not saved for historical reasons
-   // will get confused by more than one all line map
-   m_all_line_map = getMapRef("All-Line Map");
-   if (m_all_line_map == -1) {
-      // used to be called All Line Map
-      m_all_line_map = getMapRef("All Line Map");
-   }
-   if (m_all_line_map != -1) {
-      at(m_all_line_map).m_map_type = ShapeMap::ALLLINEMAP;
-   }
-
-   // VERSION_AXIAL_REGION_FIX -- this fix is now deprecated
-   // some awful things could have gone wrong in the past, but the shapemap read should fix automatically
-
    return true;
 }
 
@@ -1825,10 +648,6 @@ bool ShapeGraphs::write( std::ofstream& stream, int version, bool displayedmapon
 {
    // base class write
    ShapeMaps<ShapeGraph>::write(stream, version, displayedmaponly);
-
-   m_poly_connections.write(stream);
-   m_radial_lines.write(stream);
-
    return true;
 }
 
@@ -1988,77 +807,6 @@ void ShapeGraph::outputNet(std::ostream& netfile) const
             if (j < to) {
                netfile << (j+1) << " " << (to + 1) << " 1" << std::endl;
             }
-         }
-      }
-   }
-}
-
-/////////////////////////////////////////////////////////////////////////////////////////
-
-void ShapeGraph::makeDivisions(const prefvec<PolyConnector>& polyconnections, const pqvector<RadialLine>& radiallines, std::map<RadialKey,pvecint>& radialdivisions, std::map<int, pvecint> &axialdividers, Communicator *comm)
-{
-   time_t atime = 0;
-   if (comm) {
-      qtimer( atime, 0 );
-      comm->CommPostMessage( Communicator::NUM_RECORDS, polyconnections.size() );
-   }
-
-   for (size_t i = 0; i < polyconnections.size(); i++) {
-      PixelRefVector pixels = pixelateLine(polyconnections[i].line);
-      pvecint testedshapes;
-      auto connIter = radialdivisions.find(polyconnections[i].key);
-      size_t connindex = std::distance(radialdivisions.begin(), connIter);
-      double tolerance = sqrt(TOLERANCE_A);// * polyconnections[i].line.length();
-      for (size_t j = 0; j < pixels.size(); j++) {
-         PixelRef pix = pixels[j];
-         std::vector<ShapeRef> &shapes = m_pixel_shapes[pix.x + pix.y*m_cols];
-         for (const ShapeRef& shape: shapes) {
-            if (testedshapes.searchindex(shape.m_shape_ref) != paftl::npos) {
-               continue;
-            }
-            testedshapes.add(shape.m_shape_ref);
-            const Line& line = m_shapes.find(shape.m_shape_ref)->second.getLine();
-            //
-            if (intersect_region(line, polyconnections[i].line, tolerance * line.length()) ) {
-               switch ( intersect_line_distinguish(line, polyconnections[i].line, tolerance * line.length()) ) {
-               case 0:
-                  break;
-               case 2:
-                  {
-                     size_t index = depthmapX::findIndexFromKey(axialdividers, (int) shape.m_shape_ref);
-                     if (int(index) != shape.m_shape_ref) {
-                        throw 1; // for the code to work later this can't be true!
-                     }
-                     axialdividers[index].add(connindex);
-                     connIter->second.add(shape.m_shape_ref);
-                  }
-                  break;
-               case 1:
-                  {
-                     size_t index = depthmapX::findIndexFromKey(axialdividers, (int) shape.m_shape_ref);
-                     if (int(index) != shape.m_shape_ref) {
-                        throw 1; // for the code to work later this can't be true!
-                     }
-                     //
-                     // this makes sure actually crosses between the line and the openspace properly
-                     if (radiallines[connindex].cuts(line)) {
-                        axialdividers[index].add(connindex);
-                        connIter->second.add(shape.m_shape_ref);
-                     }
-                  }
-                  break;
-               default:
-                  break;
-               }
-            }
-         }
-      }
-      if (comm) {
-         if (qtimer( atime, 500 )) {
-            if (comm->IsCancelled()) {
-               throw Communicator::CancelledException();
-            }
-            comm->CommPostMessage( Communicator::CURRENT_RECORD, i );
          }
       }
    }
@@ -4035,130 +2783,3 @@ bool ShapeGraph::angularstepdepth(Communicator *comm)
 
    return true;
 }
-
-///////////////////////////////////////////////////////////////////////////////
-/////////////////////////////////////////////////////////////////////////////////
-
-// helper -- a little class to tidy up a set of lines
-
-void TidyLines::tidy(std::vector<Line>& lines, const QtRegion& region)
-{
-   m_region = region;
-   double maxdim = __max(m_region.width(),m_region.height());
-
-   // simple first pass -- remove very short lines
-   lines.erase(
-               std::remove_if(lines.begin(), lines.end(),
-                              [maxdim](const Line& line)
-   {return line.length() < maxdim * TOLERANCE_B;}), lines.end());
-
-   // now load up m_lines...
-   initLines(lines.size(),m_region.bottom_left,m_region.top_right);
-   for (auto& line: lines) {
-      addLine(line);
-   }
-   sortPixelLines();
-
-   std::vector<int> removelist;
-   for (size_t i = 0; i < lines.size(); i++) {
-      // n.b., as m_lines have just been made, note that what's in m_lines matches whats in lines
-      // we will use this later!
-      m_test++;
-      m_lines[i].test = m_test;
-      PixelRefVector list = pixelateLine( m_lines[i].line );
-      for (size_t a = 0; a < list.size(); a++) {
-         for (size_t b = 0; b < m_pixel_lines[ list[a].x ][ list[a].y ].size(); b++) {
-            int j = m_pixel_lines[ list[a].x ][ list[a].y ][b];
-            if (m_lines[j].test != m_test && j > (int)i && intersect_region(lines[i],lines[j],TOLERANCE_B * maxdim)) {
-               m_lines[j].test = m_test;
-               int axis_i = (lines[i].width() >= lines[i].height()) ? XAXIS : YAXIS;
-               int axis_j = (lines[j].width() >= lines[j].height()) ? XAXIS : YAXIS;
-               int axis_reverse = (axis_i == XAXIS) ? YAXIS : XAXIS;
-               if (axis_i == axis_j && fabs(lines[i].grad(axis_reverse) - lines[j].grad(axis_reverse)) < TOLERANCE_A
-                                    && fabs(lines[i].constant(axis_reverse) - lines[j].constant(axis_reverse)) < (TOLERANCE_B * maxdim)) {
-                  // check for overlap and merge
-                  int parity = (axis_i == XAXIS) ? 1 : lines[i].sign();
-                  if ((lines[i].start()[axis_i] * parity + TOLERANCE_B * maxdim) > (lines[j].start()[axis_j] * parity) &&
-                      (lines[i].start()[axis_i] * parity) < (lines[j].end()[axis_j] * parity + TOLERANCE_B * maxdim)) {
-                     int end = ((lines[i].end()[axis_i] * parity) > (lines[j].end()[axis_j] * parity)) ? i : j;
-                     lines[j].bx() = lines[end].bx();
-                     lines[j].by() = lines[end].by();
-                     removelist.push_back(i);
-                     continue; // <- don't do this any more, we've zapped it and replaced it with the later line
-                  }
-                  if ((lines[j].start()[axis_j] * parity + TOLERANCE_B * maxdim) > (lines[i].start()[axis_i] * parity) &&
-                      (lines[j].start()[axis_j] * parity) < (lines[i].end()[axis_i]  * parity + TOLERANCE_B * maxdim)) {
-                     int end = ((lines[i].end()[axis_i] * parity) > (lines[j].end()[axis_j] * parity)) ? i : j;
-                     lines[j].ax() = lines[i].ax();
-                     lines[j].ay() = lines[i].ay();
-                     lines[j].bx() = lines[end].bx();
-                     lines[j].by() = lines[end].by();
-                     removelist.push_back(i);
-                     continue; // <- don't do this any more, we've zapped it and replaced it with the later line
-                  }
-               }
-            }
-         }
-      }
-   }
-
-   // comes out sorted, remove duplicates just in case
-   removelist.erase(std::unique(removelist.begin(), removelist.end()), removelist.end());
-
-   for(auto iter = removelist.rbegin(); iter != removelist.rend(); ++iter)
-       lines.erase(lines.begin() + *iter);
-   removelist.clear();  // always clear this list, it's reused
-}
-
-void TidyLines::quicktidy(std::map<int,Line>& lines, const QtRegion& region)
-{
-   m_region = region;
-
-   double avglen = 0.0;
-
-   for (auto line: lines) {
-      avglen += line.second.length();
-   }
-   avglen /= lines.size();
-
-   double tolerance = avglen * 10e-6;
-
-   auto iter = lines.begin(), end = lines.end();
-   for(; iter != end; ) {
-       if (iter->second.length() < tolerance) {
-           iter = lines.erase(iter);
-       } else {
-           ++iter;
-       }
-   }
-
-   // now load up m_lines...
-   initLines(lines.size(),m_region.bottom_left,m_region.top_right);
-   for (auto line: lines) {
-      addLine(line.second);
-   }
-   sortPixelLines();
-
-   // and chop duplicate lines:
-   std::vector<int> removelist;
-   int i = -1;
-   for (auto line: lines) {
-      i++;
-      PixelRef start = pixelate(line.second.start());
-      for (size_t j = 0; j < m_pixel_lines[start.x][start.y].size(); j++) {
-         int k = m_pixel_lines[start.x][start.y][j];
-         if (k > (int)i && approxeq(m_lines[i].line.start(),m_lines[k].line.start(),tolerance)) {
-            if (approxeq(m_lines[i].line.end(),m_lines[k].line.end(),tolerance)) {
-               removelist.push_back(line.first);
-               break;
-            }
-         }
-      }
-   }
-   for(int remove: removelist) {
-       lines.erase(remove);
-   }
-   removelist.clear(); // always clear this list, it's reused}
-}
-
-/////////////////////////////////////////////////////
