+++ salalib/salalib.pro
@@ -43,7 +43,11 @@ SOURCES += \
     importutils.cpp \
     point.cpp \
     pafcolor.cpp \
-    spacepixfile.cpp
+    spacepixfile.cpp \
+    alllinemap.cpp \
+    axialminimiser.cpp \
+    axialpolygons.cpp \
+    tidylines.cpp
 
 HEADERS += \
     attributes.h \
@@ -82,7 +86,12 @@ HEADERS += \
     displayparams.h \
     pafcolor.h \
     options.h \
-    spacepixfile.h
+    spacepixfile.h \
+    alllinemap.h \
+    axialminimiser.h \
+    tolerances.h \
+    axialpolygons.h \
+    tidylines.h
 
 DISTFILES += \
     salascript-tests.txt
